requirements.txt
streamlit
Pillow
opencv-contrib-python
numpy
